requirements.txt
kapre==0.3.7
matplotlib==3.5.2
scikit-learn==1.1.1
SoundFile==0.10.3.post1
seaborn==0.11.2
tensorflow-gpu==2.9.1
